Extract resultClient
